README: Update the planned features section to reference the project board
Update the root README.md and remove the explicit list of features we
plan to tackle, and instead, link out to the project board that's
already tracking a good chunk of that effort. Longer term, it may make
sense to have a dedicated ROADMAP.md or a project board dedicated to
less granular tickets, highlighting the larger picture of where the
project is heading.

Signed-off-by: timflannagan <timflannagan@gmail.com>